Support dynamic overriding of chosen ClusterConfig fields
A new resource `ClusterConfigOverride` is introduced that however is not
fully standalone as its logic is heavily intertwined with
`ClusterConfig`. In the backend it is represented as a section of the
main `ClusterConfig` resource (on the same level as `spec`).

This idiosyncratic approach is chosen for the following reasons:

* To prevent race conditions and/or inconsistencies when handling
  get/set requests that could otherwise occur if the two resources were
  stored separately in the backend.
* To simply delegate to the caching mechanisms already set up for
  `ClusterConfig` and not worry about `ClusterConfigOverride` having
  to be fetched with every `ClusterConfig` get.
* To make partially applied overrides (that need to be supported not to
  reject a whole override when the main `ClusterConfig` gets updated
  with values that cause the condition for some of the original
  overrides to hold) easier to handle.
* To facilitate the presentation of the `ClusterConfig` resource,
  including investigations where it could be otherwise unclear why
  a field sticks to a certain value in spite of what's specified in
  `teleport.yaml`.

Currently, the only chosen `ClusterConfig` field that can be dynamically
overriden by means of `ClusterConfigOverride` is `session_recording`.
Its usage condition is that the override can be used to set only the
values of `off`, `node` and `node-sync`, and only if the originally
configured value is one of these as well.