Add support for organization-wide roles
This change allows to set default (catch-all) roles/logins for all users
in a github organization regardless of their teams. This is done by
allowing the following configuration in the github connector:

```
  teams_to_logins:
  - roles:
    - dev
    organization: gravitational
    team: '*'
```

Note that it is possible to assign multiple roles to a user, for example
users of the `administrators` github team can additionally be assigned
the teleport `admin` role:

```
  teams_to_logins:
  - roles:
    - admin
    organization: dev
    team: administrators
  - roles:
    - dev
    organization: gravitational
    team: '*'
```

The `MapClaims` function which returns a list of roles based on the
provided claims (for a particular github user for example), has two
nested loops. The outer loop iterates over the individual
`teams_to_logins` mappings while the inner loop aims to establish
whether the team from the current mapping is among the teams in the
provided github claims. `team: '*'` means that the `roles` of that
mapping are assigned to all users of the github org no matter what teams
they are in. Finally, note that we can stop iterating in the inner loop
once we find at least one github team that matches the team in the
current mapping.