RFD-0022: Key Agent Forwarding (#6168)
Proposes changing the default behaviour of tsh ssh -A to forward the System Key Agent to the remote machine, rather than the ephemeral Key Agent that runs inside the tsh process. This change would bring the behaviour of tsh ssh into line with the OpenSSH ssh client.

This builds on a proposed solution to #1517 in order to provide an escape hatch that allows legacy behaviour if necessary.