Variable exporting fix on AWS Terraform Guide
There was an issue with the formatting when exporting the bastion and auth IP. 

The name-value pairs must be enclosed in their own set of quotes, they should not be lumped together under a single set of quotes. 
There must be a space between the two pairs of name/value pairs. 

(I ensured I was running the latest version of AWS CLI 2 and the latest version of Terraform)