Simplifies everything
Andrew's feedback that it's OK delete the cached certificates without
checking for a role change simplifies everything. This, hopefully last,
patch makes the following changes

 - Deletes all of the role change detection code
 - Adds a parameter to ReissueUserCerts that indicates how existing
   certificates in the cache should be treated (i.e. Kept or Dropped)
 - Adds code to ProxyClient.reissueUserCerts() that takes the
   appropriate cache action.

I've added this to the ProxyClient as the cache cleanup needs to
happen _before_ the new key is added to the key store, otherwise it
becomes very difficult to differentiate between the old certificates
that need purging and new certs that need to be preserved. Given that
the rest of the system currently expects the new certs to be in the key
store before ProxyClient.reissueUserCerts() returns, this looked like
the minimal change required to add the desired behaviour.