mfa: prevent the user from deleting the last MFA device (#6585)
* mfa: prevent the user from deleting the last MFA device

When the cluster requires MFA for all users (when `second_factor` is
`on`, `u2f` or `totp`, and not `off` or `optional`), users could lock
themselves out by deleting the last device. Prevent that.

Fixes https://github.com/gravitational/teleport/issues/5803

* Make last MFA device deletion check more strict

Separate by the type of the device and which type the cluster enforces.