makefile: explicitly set SHELL to /bin/bash
When `SHELL` is not set, `make` defaults to `/bin/sh`.

On systems where `/bin/sh` is an alias for `/bin/bash`, everything works
as expected.

On systems where `/bin/sh` is actually the original Bourne Shell, some
bash-isms don't work. For example: `if [[ condition ]]` results in
`/bin/sh: 1: [[: not found`