Add workaround for Ping SAML auth requiring signing headers (#7297)
* Add workaround for PingID SAML auth

PingID's SAML auth requires `SigAlg` and `Signature` fields to be set
as query parameters during auth requests when "Enforce Signed Authn
Request" is enabled in the Ping app settings. This adds a workaround
to force gosaml2 to include these parameters when it otherwise would
not. This workaround is enabled when `provider: ping` is set in the
connector properties.

* Add `TestPingSAMLWorkaround` unit test

* Consistently refer to Ping (PingId, PingOne, etc) as just Ping

* Revert fixtures change as metadata is now embedded in `saml_test`

* Improve comments for Ping behavior workarounds

* Improve comment explanation for the `Ping` constant