Move lookup logic into `findServerByHostname`; add `export test`
This adds a new experimental subcommand that can be used to test if a
hostname is a Teleport node, which can be combined with `Match exec`
in `ssh_config` to match arbitrary nodes without requiring explicit
`Host` blocks for each cluster.

The `Server` lookup logic has been moved into a shared
`findServerByHostname` function and was optimized slightly to save a
secondary ProxyClient initialization.