Fixes racy backend test suite (#7481)
After routinely having CI builds fail in parts of the system unrelated to the change, I started to dig.

Many of these failures happened inside the backend compliance test suite, specifically in 
BackendSuite.Locking(). After a few more test failures it became apparent that the 
backend under test (etcd, memory, etc) didn't really matter; it was the test suite itself at fault.

It turns out that there are polling loops in helpers used by the test that require a clockwork.Clock
to measure a polling interval. In the test case, this was implemented by a clockwork.fakeClock
that requires manual advancing, or else the time the clock reports stays the same forever.

The working hypothesis is that, in almost all test runs, either

1. AcquireLock() succeeds on the first try, meaning it never needs to enter the polling loop, or
2. The one-off manual clock advance written into the test happened to occur during the wait, and 
   so the polling loop was woken

In either case, all would be well and the test would pass. If neither of these cases happened then the test would lock up, waiting for a clock update that would never come and eventually the test would time out after 10 minutes.

This patch:

 * Adds a 1-minute timeout to the Lock() test, so that it fails faster if the deadlock condition is hit 
 * Adds a regular, asynchronous update to the fakeClock at 10x real-time to ensure that waiting 
   code does not rely on hitting a one-of clock update, and
 * Refactors the polling loop in backend.AcquireLock() to honour context cancellation