tctl create will now ignore a blank yaml section and continue.
Added tests for `tctl create`

I broke up the Create function to separate reading the resources to be created and actually processing the stream and doing the creates. This helps in implementing the tests in a performant way vs writing out an actual temp file or mocking os.stdin which can be platform dependent.

Fixes #4703