Update DNS instructions in the AWS+EKS+Helm guide
Assuming `service/teleport` points to an ELB, which it does according to
the guide, then the change-resource-record-sets input needs to contain
an "AliasTarget" (instead of "ResourceRecords.Value").

Note that the hosted zone ID for `aws route53
change-resource-record-sets` remains as the DNS zone (ie, the zone for
"example.com"), but the hosted zone ID for the "ResourceRecordSet" is
the zone of the ELB (which doesn't necessarily match the zone for
"example.com"). I found that out by chance because I deployed my cluster
on sa-east-1 (instead of the more usual us-west or us-east).

Old change-resource-record-sets action:

```json
{
  "Action": "CREATE",
  "ResourceRecordSet": {
    "Name": "teleport.example.com",
    "Type": "A",
    "TTL": 300,
    "ResourceRecords": {
      "Value": "dualstack.a559a6696e58f48d6814781231623dc8-1207026150.sa-east-1.elb.amazonaws.com" // not allowed anymore!
    }
  }
}
```

New change-resource-record-sets action:

```json
{
  "Action": "CREATE",
  "ResourceRecordSet": {
    "Name": "teleport.example.com",
    "Type": "A",
    "AliasTarget": {
      "HostedZoneId": "<ELB_ZONE_ID>",
      "DNSName": "dualstack.a232d92df01f940339adea0e645d88bb-1576732600.us-east-1.elb.amazonaws.com",
      "EvaluateTargetHealth": false
    }
  }
}
```