Add support for a "default" role in github connector
Allow specifying team '*' in github connector. This allows specifying a
default role for all users in a github organization:

```
  teams_to_logins:
    - organization: my-github-org
      team: '*'
      logins:
        - default
```

Note that in `MapClaims` encountering a team `*`, means that the user
should receive the current mapping's claims as all users belong to the
default team `*`. Also note that we can stop iterating the inner loop
in `MapClaims` once we find at least one github team that matches the
team in the current mapping.