RDP client implementation
This is a minimal RDP client on top of the rdp-rs Rust crate.
The crate is wrapped with an FFI Rust adapter, statically compiled and
called from Go via CGO.

This PR also has toy web client to test the RDP code, and just enough of
the desktop wire protocol implementation to make it work.

I excluded the RDP client from the build via tags for now to avoid
bloating the teleport binaries.

Many more things missing that will come in later PRs, to keep this one
reviewable.