Adds prerequisites check to kube integration tests
After losing a heap of time trying to diagnose kubernetes integration
test failures only to find that they were caused by an un-primed
minikube cluster, I have added some hopefully-helpful checks to the
kubernetes integration test to assert that the tartget cluster is
ready to have test run against it.

The test failure messages point the user to a document that describes
how to initialise the cluster.