fix agent forwarding test on macOS
On macOS the default login shell is zsh, and zsh's default behavior
when redirecting to a file that already exists is to fail rather than
clobber the file contents.

Since we know the temp file will always exist and be empty, it's safe
to append (>>) - which works on both bash and zsh.