Do not deny logins in `isMFARequired` (#7739)
* Do not deny logins in `isMFARequired`

This changes `isMFARequired` to return `false` rather than an error
(e.g. not found) when a user is not authorized to access a particular
resource. This defers the authorization denial until such time as the
client actually attempts to connect.

Before Teleport v6.2, clients would immediately attempt to connect to
nodes, and if denied by RBAC rules, would print an "access denied"
error.

As of Teleport v6.2, clients first call `isMFARequired` and, if
unauthorized, abort the connection and print a "node not found"
error. As no connection is ever attempted, no audit log entry is
generated.

Fixes #7652

* Remove dead code

* Ensure users are still asked for MFA if multiple nodes match.

* Remove unnecessary debug log

* Update lib/auth/auth.go

Co-authored-by: Brian Joerger <bjoerger@gravitational.com>

* Improve clarity per review feedback

* Fix improper error log

Co-authored-by: Brian Joerger <bjoerger@gravitational.com>