Reduced shared library dependencies.
Reduced Teleport shared library dependencies on libbpf, libelf, libz.

For libbpf, switched to forked version that does not rely on "fmemopen"
which brings in a glibc 2.22 dependency. This allows binaries built on
Ubuntu 18.04 box to run on CentOS 7 as well.

For libelf and libz (which libbpf uses), the build process has been
updated to statically link both of them during the build process.