Do not exit teleport when unable to enumerate k8s cluster (#7523)
Teleport will fail to start when when a k8s cluster is unavailable when
using the kubeconfig in a `kubernetes_service` configuration. This means
that a single missing cluster can disrupt _all_ of the configured
clusters, even if the others are online.

This change makes failing the cluster credential enumeration a
per-k8s-cluster warning, rather than a stop-the-world error.

It also expands the testing shims inside the k8s proxy to allow more
sophisticated mocked scenarios, in order to test the above.

See-Also: #7215