Use github.com/google/uuid consistently
github.com/pborman/uuid is essentially a fork of the codebase by the
same author, but appears to be updated less frequently, and depends
on the Google version anyway.

With this change, we have no direct dependency on pborman/uuid,
but cannot completely shed the transitive dependency until we
update the e and reporting projects as well.

Updates #5685