RDP client implementation (#7824)
* RDP client implementation

This is a minimal RDP client on top of the rdp-rs Rust crate.
The crate is wrapped with an FFI Rust adapter, statically compiled and
called from Go via CGO.

This PR also has toy web client to test the RDP code, and just enough of
the desktop wire protocol implementation to make it work.

I excluded the RDP client from the build via tags for now to avoid
bloating the teleport binaries.

Many more things missing that will come in later PRs, to keep this one
reviewable.

* flag tweaks and updated readme for macos

* Switch to C-style strings between Go and Rust

* Use regular top-level C function instead ofthe cgo jump function

* Fix bitmap buffer memory leak

* Consistent naming for CGO types

* Pass rust object reference to Go

* Clean up FFI error string management

* Extract a thin C wrapper for read_rdp_output

* Use the log crate in rust

* Small rust cleanups

* Fix shellcheck nit in run.sh

* Fix RDP client memory release

* Allow Rust code to compile on any unix

* Force Alpha channel to 100% always

For some reason, on Windows 10 with bitmap decompression the Alpha
always ends up as 0. This makes everything transparent.

* Implement screen size and credential negotiation

* desktop protocol: remove password prompt

It was decided that supporting passwords natively is a bad product
decision. We will only support certificate-based authn and only in
ActiveDirectory environments.

Until we implement certificate support, passwords can be injected via an
environment variable for testing. It will be removed before the beta
release.

* Address review feedback

Co-authored-by: Isaiah Becker-Mayer <isaiah@goteleport.com>