Fix node registration backwards compatibility
With the consolidation of PackedKeys and proto.Certs in #8193,
I mistakenly changed the JSON that is returned from the /tokens/register
endpoint and broke the ability for older clients (7.x and earlier)
to register with newer (8.x and up) auth servers.

This wasn't caught by the type system as part of the refactor because
our API server handlers return interface{} and simply JSON encode whatever
is returned.

This commit preserves the original wire format by introducing an
unexported type with the original JSON keys and converting back
to proto.Certs before/after putting data on the wire.