Specify platform when building our buildbox
On Apple silicon, Docker will default to building arm64 images.
Some of the packages our image tries to install are only available
for amd64, so the build cannot complete.

This should have no impact for devs running on amd64, and makes it
possible for devs on Apple silion to run build tasks that require
the buildbox (generating protos, for example)