Stop creating a default admin role
With this change, we no longer create a default admin role as it
is too powerful. Instead we prefer the "preset" edit, audit, and
access roles.

Also add wildcard labels for Windows desktop access to the preset
access role. Note: this only applies for new clusters - teleport
will not overwrite the access role on existing clusters to avoid
clobbering user modifications.