Ensure there are no '.' characters in dynamic desktop names
We embed the name of the desktop in the server name when establishing
a connection, and the backend uses this to figure out which desktop
to route to.

Our wildcard certificate for desktop access is only valid for
*.desktop.teleport.cluster.local, so we need replace '.' characters
in the name in order to avoid creating a new subdomain for which
the cert is not valid.

This logic already exists for statically defined desktops, and
was lacking for desktops discovered via LDAP.