Make LDAP desktop discovery disabled by default
Since our LDAP-based desktop discovery is not very configurable,
we opt to have it disabled by default.

Teleport will log a warning if desktop discovery is disabled and there
are no statically defined Windows desktops. In this case, the Windows
Desktop Service will simply sit idle, as there will be no desktops
available to connect to.

This commit implements the above, but also paves the way towards
a more flexible discovery system (described in the RFD).

We introduce a new config section:

    discovery:
      base_dn: '*'
      filters:
      - filter1
      - filter2

For now, the only valid value for base_dn is the wildcard, which
instructs Teleport to search from the domain root. Additionally,
teleport will validate that any provided filters are valid but
does not currently respect them when performing the search.

Future updates will allow for changing the base DN to something
more specific and filtering the results with LDAP filters.